refactors invalid JSON aop test
