+++ src/test/java/com/lifu/wsms/reload/exception/RestExceptionHandlerTest.java
@@ -1,5 +1,6 @@
 package com.lifu.wsms.reload.exception;
 
+import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -8,14 +9,15 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import static com.lifu.wsms.reload.controller.student.StudentController.STUDENT_PATH;
+import static com.lifu.wsms.reload.util.TestUtil.getCreateStudentRequestJsonString;
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @ExtendWith(MockitoExtension.class)
 @SpringBootTest
@@ -31,44 +33,7 @@ class RestExceptionHandlerTest {
 
     @Test
     void handleHttpMessageNotReadableException() throws Exception {
-        var request = "{\n" +
-                "    \"studentId\": \"KSK-2024-1234\",\n" +
-                "    \"firstName\": \"David\",\n" +
-                "    \"middleName\": \"Owogoga\",\n" +
-                "    \"lastName\": \"Lifu\",\n" +
-                "    \"dob\": \"2010-01-01\",\n" +
-                "    \"gender\": \"male lion\",\n" +
-                "    \"address\": {\n" +
-                "        \"houseNumber\": \"21\",\n" +
-                "        \"streetName\": \"Lyon Crescent\",\n" +
-                "        \"area\": \"Stirling\",\n" +
-                "        \"localGovtArea\": \"Stirling\",\n" +
-                "        \"state\": \"Scotland\",\n" +
-                "        \"country\": \"United Kingdom\"\n" +
-                "    },\n" +
-                "    \"contact\": {\n" +
-                "        \"email\": \"davidlifu@gmail.com\",\n" +
-                "        \"mobilePhone\": \"07766433489\",\n" +
-                "        \"telephone\": \"\"\n" +
-                "    },\n" +
-                "    \"legalGuardian\": {\n" +
-                "        \"isBiologicalParentListed\": true,\n" +
-                "        \"father\": \"Anthony Lifu\",\n" +
-                "        \"fatherContactInformation\": {\n" +
-                "            \"email\": \"tlifu75@gmail.com\",\n" +
-                "            \"mobilePhone\": \"07766433489\",\n" +
-                "            \"telephone\": \"\"\n" +
-                "        },\n" +
-                "        \"mother\": \"\",\n" +
-                "        \"motherContactInformation\": {\n" +
-                "            \"email\": \"ladilifu@gmail.com\",\n" +
-                "            \"mobilePhone\": \"07766433489\",\n" +
-                "            \"telephone\": \"\"\n" +
-                "        }\n" +
-                "    },\n" +
-                "    \"currentGrade\": \"SSS120241\",\n" +
-                "    \"isDisabled\": false\n" +
-                "}";
+        var request = getCreateStudentRequestJsonString();
 
         mockMvc.perform(
                         post(STUDENT_PATH)
@@ -76,6 +41,11 @@ class RestExceptionHandlerTest {
                                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                                 .content(request))
                 .andExpect(status().isBadRequest())
-                .andExpect(content().string(containsString("Invalid JSON request body")));
+                .andExpect(content().string(containsString("Invalid JSON request body")))
+                .andExpect(jsonPath("$.apiResponse.responseCode").value("407"))
+                .andExpect(jsonPath("$.apiResponse.responseMessage").value("Invalid JSON request body"))
+                .andExpect(jsonPath("$.apiResponse.httpStatusCode").value("BAD_REQUEST"))
+                .andExpect(jsonPath("$.apiResponse.error").value(true));
     }
+
 }
